refactor: MemberApiController, MemberService 수정 및 MemberUpateParam 추가

+++ src/main/java/com/board/web/MemberApiController.java
@@ -1,5 +1,6 @@
 package com.board.web;
 
+
 import com.board.service.MemberService;
 import com.board.web.dto.member.MemberListResponseDto;
 import com.board.web.dto.member.MemberResponseDto;
@@ -11,6 +12,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+
 @Api("Member Controller API V1")
 @RequiredArgsConstructor
 @RestController
@@ -19,38 +22,48 @@ public class MemberApiController {
     private final MemberService service;
 
 
-    @GetMapping("api/v1/members")
-    @ApiOperation(value = "회원 목록", notes = "회원 전체 목록으로 보여준다")
-    public List<MemberListResponseDto> findAll() {
-        List<MemberListResponseDto> all = service.findAll();
-        return all;
+    @PostMapping("/api/v1/member")
+    @ApiOperation(value = "회원 등록", notes = "회원을 등록한다")
+    public Long save(@RequestBody MemberSaveRequestDto request) {
+        return service.save(request.toParam());
     }
 
     @GetMapping("/api/v1/member/{no}")
-    @ApiOperation(value = "회원 조회", notes = "회원 한명을 보여준다 ")
-    public MemberResponseDto findById(@PathVariable Long no) {
+    @ApiOperation(value = "회원 조회", notes = "회원 한명을 조회한다. ")
+    public MemberResponseDto findMember(@PathVariable Long no){
+        if(service.findByNo(no).equals(null)) {
+//            throw new IllegalArgumentException("해당 번호의 회원이 없습니다");
+            throw new IllegalArgumentException("해당 번호의 회원이 없습니다");
+        }
         return service.findByNo(no);
     }
 
     @PutMapping("api/v1/member/{no}")
     @ApiOperation(value = "회원 수정", notes = "회원 한명의 정보를 수정한다 ")
-    public Long update(@PathVariable Long no, @RequestBody MemberUpdateRequestDto dto) {
-        return service.update(no, dto);
-
+    public void update(@PathVariable Long no,
+                       @RequestBody MemberUpdateRequestDto request) {
+        if(service.findByNo(no).equals(null)) {
+            throw new IllegalArgumentException("해당 번호의 회원이 없습니다");
+        }
+        service.update(no, request.toParam());
+    }
+    @GetMapping("api/v1/members")
+    @ApiOperation(value = "회원 목록", notes = "회원 전체 목록으로 보여준다")
+    public List<MemberListResponseDto> findAll() {
+        return service.findAll();
     }
 
     @DeleteMapping("/api/v1/member/{no}")
     @ApiOperation(value = "회원 삭제", notes = "해당 회원을 삭제한다")
     public Long delete(@PathVariable Long no) {
+        if(service.findByNo(no).equals(null)) {
+            throw new IllegalArgumentException("해당 번호의 회원이 없습니다");
+        }
         service.delete(no);
         return no;
     }
 
-    @PostMapping("/api/v1/member")
-    @ApiOperation(value = "회원 등록", notes = "회원을 등록한다")
-    public Long save(@RequestBody MemberSaveRequestDto dto) {
-        return service.save(dto);
-    }
+
